requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.12
pandas>=1.3.0
Pillow>=8.3.1
tqdm>=4.62.0
